Them chinh sua tai khoan ca nhan

+++ src/main/java/qbit/entier/user_service/service/UserService.java
@@ -12,6 +12,7 @@ import org.springframework.web.multipart.MultipartFile;
 import qbit.entier.user_service.client.AuthServiceClient;
 import qbit.entier.user_service.dto.AccountDto;
 import qbit.entier.user_service.dto.UserAccountDto;
+import qbit.entier.user_service.dto.UserDto;
 import qbit.entier.user_service.entity.User;
 import qbit.entier.user_service.repository.UserRepository;
 import qbit.entier.user_service.util.FileUtil;
@@ -57,6 +58,27 @@ public class UserService {
         return userRepository.findUserByAccountId(auth.getId()).map((user -> UserAccountDto.fromEntity(user, auth)));
     }
 
+    public UserDto updateSelf(User user) {
+        UserAccountDto userAccountDto = getSelf().orElseThrow(() ->
+                new EntityNotFoundException("Not found"));
+
+        User updatedUser = userRepository.findById(userAccountDto.getId()).orElseThrow(
+                () -> new EntityNotFoundException("Not found")
+        );
+
+        if(user.getEmail() != null)
+            updatedUser.setEmail(user.getEmail());
+        if(user.getPhone() != null)
+            updatedUser.setPhone(user.getPhone());
+        if(user.getFullName() != null)
+            updatedUser.setFullName(user.getFullName());
+        if(user.getIsMale() != null)
+            updatedUser.setIsMale(user.getIsMale());
+
+        return UserDto.fromEntity(userRepository.save(updatedUser));
+
+    }
+
     public Optional<User> getSelfUser() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         AccountDto auth = (AccountDto) authentication.getPrincipal();
@@ -92,7 +114,7 @@ public class UserService {
 
         if (optionalUser.isPresent()) {
             User user = optionalUser.get();
-            if (user.getAvatar() != null) {
+            if (user.getAvatar() != null && fileUtil.fileExists(user.getAvatar())) {
                 fileUtil.deleteFile(user.getAvatar());
             }
             String avatarPath = fileUtil.saveFile(file);
@@ -145,8 +167,9 @@ public class UserService {
             throw new EntityNotFoundException("Failed to retrieve account information from auth service");
         }
 
+        if(existingUser.getAvatar() != null && fileUtil.fileExists(existingUser.getAvatar()))
+            fileUtil.deleteFile(existingUser.getAvatar());
+
         userRepository.delete(existingUser);
     }
-
-
 }
